+++ src/client/src/pages/settings/dataLoader.jsx
@@ -1,22 +1,13 @@
 import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { CircularProgress } from "@mui/material";
+import { useDispatch } from "react-redux";
 import { loadSettings } from "../../api-lib/index";
-import { SplashScreen } from "../../auth/SplashScreen";
 
 export const DataLoader = ({ children }) => {
   const dispatch = useDispatch();
-  const isReady = useSelector(state => state.dataLoaderState.isReady);
 
   useEffect(() => {
     dispatch(loadSettings());
   }, [dispatch]);
 
-  return isReady ? (
-    children
-  ) : (
-    <SplashScreen>
-      <CircularProgress />
-    </SplashScreen>
-  );
+  return children;
 };